+++ Changelog.md
@@ -9,6 +9,7 @@
 - Add settings to Nametags to adjust text sizes
 - UB expression patch now properly patches VTOptGet and VTOptSet meta actions
 - Add Nametags for allegiance members now have a unique color
+- Add `getworldname[]` expression to return the name of the currently connected world / server
 
 ## 0.2.1 (2021-01-02) [UtilityBeltInstaller-0.2.1.exe](https://gitlab.com/utilitybelt/utilitybelt.gitlab.io/uploads/8552636461a1e2d1ec0c8c140cc7c9dd/UtilityBeltInstaller-0.2.1.exe)
 - Disabled character options profiles -- too buggy